+++ app/src/main/java/com/homemade/anothertodo/db/dao/SingleTaskDao.kt
@@ -23,7 +23,7 @@ interface SingleTaskDao {
     suspend fun deleteAll()
 
     @Query("SELECT * FROM single_task_table WHERE id = :id")
-    fun getTask(id: Int): SingleTask?
+    fun getTask(id: Long): SingleTask?
 
     @Query("SELECT * FROM single_task_table ORDER BY name ASC")
     fun getTasks(): Flow<List<SingleTask>>
